Tutorial: fix numpy deprecation warning
Numpy 1.26.2 now complains about casting a Python integer to a numpy
dtype for which it is out of range. Work around that by first converting
to np.int_.

+++ examples/tut_4_send_perf.py
@@ -54,7 +54,7 @@ def main():
     start = time.perf_counter()
     for i in range(n_heaps):
         item_group["timestamp"].value = i * heap_size
-        item_group["adc_samples"].value = np.full(heap_size, i, np.int8)
+        item_group["adc_samples"].value = np.full(heap_size, np.int_(i), np.int8)
         heap = item_group.get_heap()
         stream.send_heap(heap)
     elapsed = time.perf_counter() - start
